+++ src/screens/Home/Audience.js
@@ -1,0 +1,9 @@
+import React from "react";
+import { View, TouchableOpacity, Text, Image } from "react-native";
+import { useNavigation } from "@react-navigation/native";
+
+const Audience = () => {
+  return <View></View>;
+};
+
+export default Audience;