+++ CHANGELOG.md
@@ -9,6 +9,7 @@ NOTE: all lowercase and no ending punctuation marks
 
 ## Next release v1.0
 
+- [CY] - [#9](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/9) app(feature): home bottom tabs navigation
 - [CY] - [#7](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/7) app(feature): add dependencies to enable expo to publish the app online
 - [CY] - [#5](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/5) app(feature): base project
 - [CY] - [#3](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/3) markdown(clean): clean and update markdowns